Fixed unimplemented method since change in faceness calculation
As `get_score` no longer returns `faceness` (error in calculation; see 3a17220), it does not make sense to calculate the faceness of an image using a single feature, so we have ensured this is not implemented.  We have changed the tests accordingly.  You should use the other method of `get_faceness`, which calculates the faceness given potentially many classifiers.