Increase Julia compat version to 1.4
While I would like to support older versions of Julia, currently this
will be a little bit of work that I don't want to have to do at the
current point in time.  Increasing Julia compat version to version 1.4
in the interest of `readdir`'s keyword arguments `join` and `sort`,
the former of which is used in src/Utils.jl methods for training a
Voila Jones model.  It wouldn't be hard to implement `join` manually,
but I know not what else will need fixing in older versions of Julia,
and I don't even know if people use this package, let alone use this
package with Julia version < 1.4.  Perhaps upon request I will support
Julia versions < 1.4, but for now, this compat version will do.